+++ tests/migrations/000000-000001-migration.php
@@ -4,4 +4,4 @@ declare(strict_types=1);
 
 use Celemas\Quma\Tests\Util\TestMigration1;
 
-return new TestMigration1();
+return TestMigration1::class;